requirements.txt
beautifulsoup4==4.9.3
Brotli==1.0.7
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dash==1.12.0
dash-core-components==1.10.0
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Compress==1.5.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.3
future==0.18.2
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.22.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
lxml==4.6.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
mysqlclient==2.0.1
numpy==1.19.2
openpyxl==3.0.5
pandas==1.1.3
plotly==4.8.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
rdbtools==0.1.15
redis==3.5.3
requests==2.24.0
retrying==1.3.3
rsa==4.6
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
uritemplate==3.0.1
urllib3==1.25.11
Werkzeug==1.0.1
